feat(app): 周次视图接入设置页、挂件⋯菜单与托盘
- WidgetSettings.WeekView（默认 Current，字符串落盘：转换器只挂枚举类型，不动全局 STJ 选项）
- AppStartupOptions：--today YYYY-MM-DD（今日唯一真源）+ --week-view all|current|odd|even（只影响本次运行）
- MainWindow：WeekViewEnabled 有效值喂 BoardOptions；视图一变整树重排 + 落盘；刷新加 [weekview] ASCII 锚点（view/today/week/todaySessions/hidden/menu）
- 设置「外观」页加「周次视图」ComboBox（说明静态写明开学日未知时显示全部周次）
- ⋯ 菜单加「周次视图」子菜单（四项 RadioMenuFlyoutItem，选中态读有效值）
- 托盘加同名子菜单（普通勾 MF_CHECKED，不用 MFT_RADIOCHECK）：TrayMenuItem.Children + MF_POPUP，AppendMenuW 第三参改 nint（UINT_PTR，写 uint 会截断子菜单句柄）
- VerifySmoke 按有效视图复算可见时段数，[smoke] 行补 view/today/week

+++ dotnet/Tjt.App/Win32/NativeMethods.cs
@@ -383,9 +383,13 @@ internal static partial class NativeMethods
     [DllImport("user32.dll")]
     internal static extern nint CreatePopupMenu();
 
-    /// <summary>追加菜单项。</summary>
+    /// <summary>
+    /// 追加菜单项。第三个参数原生类型是 <c>UINT_PTR</c>：普通项放命令 id，
+    /// <c>MF_POPUP</c> 时放**子菜单句柄** —— 句柄在 x64 上可能超过 32 位，所以必须声明成
+    /// <see cref="nint"/>（写成 <c>uint</c> 会截断句柄，子菜单整个失效）。
+    /// </summary>
     [DllImport("user32.dll", EntryPoint = "AppendMenuW", CharSet = CharSet.Unicode)]
-    internal static extern bool AppendMenu(nint menu, uint flags, uint id, string? text);
+    internal static extern bool AppendMenu(nint menu, uint flags, nint id, string? text);
 
     /// <summary>弹出菜单（<c>TPM_RETURNCMD</c> 时返回选中的命令 id）。</summary>
     [DllImport("user32.dll")]
